requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Werkzeug==3.0.1
email-validator==2.1.0
Pillow>=10.0.0
stripe>=8.0.0
paypalrestsdk>=1.13.0
python-dotenv>=1.0.0
python-slugify>=8.0.0
gunicorn>=21.0.0
rembg==2.0.59
requests>=2.31.0
twilio>=8.0.0